private化(一部、残りは週末にでも)

+++ include/internal/battle/GameCardMono.hpp
@@ -5,6 +5,6 @@ using namespace std;
 class GameCardMono{
 public:
     string cardName;
-    int cardLevel;
-    int needEXP; //制限回数を超えたらここが-1になる
+    short cardLevel;
+    short needEXP; //制限回数を超えたらここが-1になる
 };
